add initial support for shared versions
Not backward compatible with previous kernels, so please only use with
kernel that has version table support (this will need to be cleaned up).

No READDIRPLUS support in the kernel for versioned entries, so disable for
now.

Attribute timeout is set to "infinity", so changes to underlying filesystem
won't be visible.  This also needs to be fixed, but is best for testing the
versioning since the shared version is the only thing that will force
refreshing metadata and dcache lookups.

No caching metadata modifications yet.

Start "ireg" daemon before starting any fuse servers.

+++ example/passthrough_ll.c
@@ -51,11 +51,17 @@
 #include <assert.h>
 #include <errno.h>
 #include <err.h>
+#include <semaphore.h>
 #include <inttypes.h>
 #include <pthread.h>
 #include <sys/syscall.h>
 #include <sys/file.h>
 #include <sys/xattr.h>
+#include <sys/mman.h>
+#include <sys/socket.h>
+#include <sys/un.h>
+
+#include "ireg.h"
 
 /* We are re-using pointers to our `struct lo_inode` and `struct
    lo_dirp` elements as inodes. This means that we must be able to
@@ -78,6 +84,8 @@ struct lo_inode {
 	ino_t ino;
 	dev_t dev;
 	uint64_t refcount; /* protected by lo->mutex */
+	uint64_t version_offset;
+	uint64_t ireg_refid;
 };
 
 struct lo_cred {
@@ -88,6 +96,7 @@ struct lo_cred {
 enum {
 	CACHE_NEVER,
 	CACHE_NORMAL,
+	CACHE_SHARED,
 	CACHE_ALWAYS,
 };
 
@@ -104,6 +113,9 @@ struct lo_data {
 	int readdirplus_set;
 	int readdirplus_clear;
 	struct lo_inode root; /* protected by lo->mutex */
+	int ireg_sock;
+	int64_t *version_table;
+	uint64_t version_table_size;
 };
 
 static const struct fuse_opt lo_opts[] = {
@@ -129,6 +141,8 @@ static const struct fuse_opt lo_opts[] = {
 	  offsetof(struct lo_data, cache), CACHE_NEVER },
 	{ "cache=auto",
 	  offsetof(struct lo_data, cache), CACHE_NORMAL },
+	{ "cache=shared",
+	  offsetof(struct lo_data, cache), CACHE_SHARED },
 	{ "cache=always",
 	  offsetof(struct lo_data, cache), CACHE_ALWAYS },
 	{ "readdirplus",
@@ -137,6 +151,7 @@ static const struct fuse_opt lo_opts[] = {
 	  offsetof(struct lo_data, readdirplus_clear), 1 },
 	FUSE_OPT_END
 };
+static void unref_inode(struct lo_data *lo, struct lo_inode *inode, uint64_t n);
 
 static struct lo_data *lo_data(fuse_req_t req)
 {
@@ -180,13 +195,35 @@ static void lo_init(void *userdata,
 			fprintf(stderr, "lo_init: activating flock locks\n");
 		conn->want |= FUSE_CAP_FLOCK_LOCKS;
 	}
+	/* TODO: shared version support for readdirplus */
 	if ((lo->cache == CACHE_NEVER && !lo->readdirplus_set) ||
-	    lo->readdirplus_clear) {
+	    lo->readdirplus_clear || lo->cache == CACHE_SHARED) {
 		if (lo->debug)
 			fprintf(stderr, "lo_init: disabling readdirplus\n");
 		conn->want &= ~FUSE_CAP_READDIRPLUS;
 	}
 
+}
+
+static int64_t *version_ptr(struct lo_data *lo, struct lo_inode *inode)
+{
+	return lo->version_table + inode->version_offset;
+}
+
+static int64_t get_version(struct lo_data *lo, struct lo_inode *inode)
+{
+	if (!inode->version_offset)
+		return 0;
+
+	return __atomic_load_8(version_ptr(lo, inode), __ATOMIC_SEQ_CST);
+}
+
+static void update_version(struct lo_data *lo, struct lo_inode *inode)
+{
+	if (!inode->version_offset)
+		return;
+
+	__atomic_add_fetch(version_ptr(lo, inode), 1, __ATOMIC_SEQ_CST);
 }
 
 static void lo_getattr(fuse_req_t req, fuse_ino_t ino,
@@ -232,6 +269,7 @@ static void lo_setattr(fuse_req_t req, fuse_ino_t ino, struct stat *attr,
 	char procname[64];
 	struct lo_inode *inode = lo_inode(req, ino);
 	int ifd = inode->fd;
+	struct lo_data *lo = lo_data(req);
 	int res;
 
 	if (valid & FUSE_SET_ATTR_MODE) {
@@ -290,6 +328,7 @@ static void lo_setattr(fuse_req_t req, fuse_ino_t ino, struct stat *attr,
 		if (res == -1)
 			goto out_err;
 	}
+	update_version(lo, inode);
 
 	return lo_getattr(req, ino, fi);
 
@@ -316,6 +355,67 @@ static struct lo_inode *lo_find(struct lo_data *lo, struct stat *st)
 	return ret;
 }
 
+struct msgreply {
+	struct lo_inode *inode;
+	sem_t ready;
+};
+
+static void get_shared(struct lo_data *lo, struct lo_inode *inode)
+{
+	int res;
+	struct msgreply rep = {
+		.inode = inode,
+	};
+	struct ireg_msg msg = {
+		.op = IREG_GET,
+		.handle = (uintptr_t) &rep,
+		.get = {
+			.ino = inode->ino,
+			.dev = inode->dev,
+		},
+	};
+
+	if (lo->ireg_sock == -1) {
+		inode->version_offset = 0;
+		return;
+	}
+
+	sem_init(&rep.ready, 0, 0);
+
+	res = write(lo->ireg_sock, &msg, sizeof(msg));
+	if (res != sizeof(msg)) {
+		if (res == -1)
+			warn("write(lo->ireg_sock, {IREG_GET, ...})");
+		else
+			warnx("short write to ireg_sock: %i", res);
+		return;
+	}
+
+	while (sem_wait(&rep.ready));
+	sem_destroy(&rep.ready);
+}
+
+static void put_shared(struct lo_data *lo, struct lo_inode *inode)
+{
+	int res;
+	struct ireg_msg msg = {
+		.op = IREG_PUT,
+		.put.refid = inode->ireg_refid,
+	};
+
+	if (lo->ireg_sock == -1)
+		return;
+
+	res = write(lo->ireg_sock, &msg, sizeof(msg));
+	if (res != sizeof(msg)) {
+		if (res == -1)
+			warn("write(lo->ireg_sock, {IREG_PUT, ...})");
+		else
+			warnx("short write to ireg_sock: %i", res);
+		return;
+	}
+}
+
 static int lo_do_lookup(fuse_req_t req, fuse_ino_t parent, const char *name,
 			 struct fuse_entry_param *e)
 {
@@ -323,13 +423,13 @@ static int lo_do_lookup(fuse_req_t req, fuse_ino_t parent, const char *name,
 	int res;
 	int saverr;
 	struct lo_data *lo = lo_data(req);
-	struct lo_inode *inode;
+	struct lo_inode *inode, *dir = lo_inode(req, parent);
 
 	memset(e, 0, sizeof(*e));
 	e->attr_timeout = lo->timeout;
 	e->entry_timeout = lo->timeout;
 
-	newfd = openat(lo_fd(req, parent), name, O_PATH | O_NOFOLLOW);
+	newfd = openat(dir->fd, name, O_PATH | O_NOFOLLOW);
 	if (newfd == -1)
 		goto out_err;
 
@@ -337,7 +437,7 @@ static int lo_do_lookup(fuse_req_t req, fuse_ino_t parent, const char *name,
 	if (res == -1)
 		goto out_err;
 
-	inode = lo_find(lo_data(req), &e->attr);
+	inode = lo_find(lo, &e->attr);
 	if (inode) {
 		close(newfd);
 		newfd = -1;
@@ -352,8 +452,11 @@ static int lo_do_lookup(fuse_req_t req, fuse_ino_t parent, const char *name,
 		inode->is_symlink = S_ISLNK(e->attr.st_mode);
 		inode->refcount = 1;
 		inode->fd = newfd;
+		newfd = -1;
 		inode->ino = e->attr.st_ino;
 		inode->dev = e->attr.st_dev;
+
+		get_shared(lo, inode);
 
 		pthread_mutex_lock(&lo->mutex);
 		prev = &lo->root;
@@ -364,11 +467,24 @@ static int lo_do_lookup(fuse_req_t req, fuse_ino_t parent, const char *name,
 		prev->next = inode;
 		pthread_mutex_unlock(&lo->mutex);
 	}
+
+	e->initial_version = get_version(lo, inode);
+	res = fstatat(inode->fd, "", &e->attr,
+		      AT_EMPTY_PATH | AT_SYMLINK_NOFOLLOW);
+	if (res == -1) {
+		unref_inode(lo, inode, 1);
+		goto out_err;
+	}
+
 	e->ino = (uintptr_t) inode;
+	e->version_offset = inode->version_offset;
 
 	if (lo_debug(req))
-		fprintf(stderr, "  %lli/%s -> %lli\n",
-			(unsigned long long) parent, name, (unsigned long long) e->ino);
+		fprintf(stderr, "  %lli/%s -> %lli (version_table[%lli]=%lli)\n",
+			(unsigned long long) parent, name,
+			(unsigned long long) e->ino,
+			(unsigned long long) e->version_offset,
+			(unsigned long long) e->initial_version);
 
 	return 0;
 
@@ -447,6 +563,7 @@ static void lo_mknod_symlink(fuse_req_t req, fuse_ino_t parent,
 	struct lo_inode *dir = lo_inode(req, parent);
 	struct fuse_entry_param e;
 	struct lo_cred old = {};
+	struct lo_data *lo = lo_data(req);
 
 	saverr = ENOMEM;
 	inode = calloc(1, sizeof(struct lo_inode));
@@ -469,6 +586,8 @@ static void lo_mknod_symlink(fuse_req_t req, fuse_ino_t parent,
 
 	if (res == -1)
 		goto out;
+
+	update_version(lo, lo_inode(req, parent));
 
 	saverr = lo_do_lookup(req, parent, name, &e);
 	if (saverr)
@@ -551,6 +670,8 @@ static void lo_link(fuse_req_t req, fuse_ino_t ino, fuse_ino_t parent,
 	inode->refcount++;
 	pthread_mutex_unlock(&lo->mutex);
 	e.ino = (uintptr_t) inode;
+	update_version(lo, inode);
+	update_version(lo, lo_inode(req, parent));
 
 	if (lo_debug(req))
 		fprintf(stderr, "  %lli/%s -> %lli\n",
@@ -565,13 +686,39 @@ out_err:
 	fuse_reply_err(req, saverr);
 }
 
+static struct lo_inode *lookup_name(fuse_req_t req, fuse_ino_t parent,
+				    const char *name)
+{
+	int res;
+	struct stat attr;
+
+	res = fstatat(lo_fd(req, parent), name, &attr,
+		      AT_EMPTY_PATH | AT_SYMLINK_NOFOLLOW);
+	if (res == -1)
+		return NULL;
+
+	return lo_find(lo_data(req), &attr);
+}
+
 static void lo_rmdir(fuse_req_t req, fuse_ino_t parent, const char *name)
 {
 	int res;
+	struct lo_inode *inode = lookup_name(req, parent, name);
+	struct lo_data *lo = lo_data(req);
+
+	if (!inode) {
+		fuse_reply_err(req, EIO);
+		return;
+	}
 
 	res = unlinkat(lo_fd(req, parent), name, AT_REMOVEDIR);
-
-	fuse_reply_err(req, res == -1 ? errno : 0);
+	if (res == -1) {
+		fuse_reply_err(req, errno);
+	} else {
+		update_version(lo, inode);
+		update_version(lo, lo_inode(req, parent));
+		fuse_reply_err(req, 0);
+	}
 }
 
 static void lo_rename(fuse_req_t req, fuse_ino_t parent, const char *name,
@@ -579,6 +726,15 @@ static void lo_rename(fuse_req_t req, fuse_ino_t parent, const char *name,
 		      unsigned int flags)
 {
 	int res;
+	struct lo_inode *oldinode = lookup_name(req, parent, name);
+	struct lo_inode *newinode = lookup_name(req, newparent, newname);
+	struct lo_data *lo = lo_data(req);
+
+	if (!oldinode) {
+		fuse_reply_err(req, EIO);
+		return;
+	}
+
 
 	if (flags) {
 #ifndef SYS_renameat2
@@ -596,17 +752,37 @@ static void lo_rename(fuse_req_t req, fuse_ino_t parent, const char *name,
 
 	res = renameat(lo_fd(req, parent), name,
 			lo_fd(req, newparent), newname);
-
-	fuse_reply_err(req, res == -1 ? errno : 0);
+	if (res == -1) {
+		fuse_reply_err(req, errno);
+	} else {
+		update_version(lo, oldinode);
+		if (newinode)
+			update_version(lo, newinode);
+		update_version(lo, lo_inode(req, parent));
+		update_version(lo, lo_inode(req, newparent));
+		fuse_reply_err(req, 0);
+	}
 }
 
 static void lo_unlink(fuse_req_t req, fuse_ino_t parent, const char *name)
 {
 	int res;
+	struct lo_inode *inode = lookup_name(req, parent, name);
+	struct lo_data *lo = lo_data(req);
+
+	if (!inode) {
+		fuse_reply_err(req, EIO);
+		return;
+	}
 
 	res = unlinkat(lo_fd(req, parent), name, 0);
-
-	fuse_reply_err(req, res == -1 ? errno : 0);
+	if (res == -1) {
+		fuse_reply_err(req, errno);
+	} else {
+		update_version(lo, inode);
+		update_version(lo, lo_inode(req, parent));
+		fuse_reply_err(req, 0);
+	}
 }
 
 static void unref_inode(struct lo_data *lo, struct lo_inode *inode, uint64_t n)
@@ -627,6 +803,7 @@ static void unref_inode(struct lo_data *lo, struct lo_inode *inode, uint64_t n)
 
 		pthread_mutex_unlock(&lo->mutex);
 		close(inode->fd);
+		put_shared(lo, inode);
 		free(inode);
 
 	} else {
@@ -857,6 +1034,7 @@ static void lo_create(fuse_req_t req, fuse_ino_t parent, const char *name,
 	struct fuse_entry_param e;
 	int err;
 	struct lo_cred old = {};
+	struct lo_data *lo = lo_data(req);
 
 	if (lo_debug(req))
 		fprintf(stderr, "lo_create(parent=%" PRIu64 ", name=%s)\n",
@@ -872,6 +1050,8 @@ static void lo_create(fuse_req_t req, fuse_ino_t parent, const char *name,
 	lo_restore_cred(&old);
 
 	if (!err) {
+		update_version(lo, lo_inode(req, parent));
+
 		fi->fh = fd;
 		err = lo_do_lookup(req, parent, name, &e);
 	}
@@ -992,6 +1172,7 @@ static void lo_write_buf(fuse_req_t req, fuse_ino_t ino,
 	(void) ino;
 	ssize_t res;
 	struct fuse_bufvec out_buf = FUSE_BUFVEC_INIT(fuse_buf_size(in_buf));
+	struct lo_data *lo = lo_data(req);
 
 	out_buf.buf[0].flags = FUSE_BUF_IS_FD | FUSE_BUF_FD_SEEK;
 	out_buf.buf[0].fd = fi->fh;
@@ -1002,10 +1183,12 @@ static void lo_write_buf(fuse_req_t req, fuse_ino_t ino,
 			ino, out_buf.buf[0].size, (unsigned long) off);
 
 	res = fuse_buf_copy(&out_buf, in_buf, 0);
-	if(res < 0)
+	if(res < 0) {
 		fuse_reply_err(req, -res);
-	else
+	} else {
+		update_version(lo, lo_inode(req, ino));
 		fuse_reply_write(req, (size_t) res);
+	}
 }
 
 static void lo_statfs(fuse_req_t req, fuse_ino_t ino)
@@ -1024,6 +1207,8 @@ static void lo_fallocate(fuse_req_t req, fuse_ino_t ino, int mode,
 			 off_t offset, off_t length, struct fuse_file_info *fi)
 {
 	int err;
+	struct lo_data *lo = lo_data(req);
+
 	(void) ino;
 
 	if (mode) {
@@ -1032,6 +1217,8 @@ static void lo_fallocate(fuse_req_t req, fuse_ino_t ino, int mode,
 	}
 
 	err = posix_fallocate(fi->fh, offset, length);
+	if (!err)
+		update_version(lo, lo_inode(req, ino));
 
 	fuse_reply_err(req, err);
 }
@@ -1164,6 +1351,7 @@ static void lo_setxattr(fuse_req_t req, fuse_ino_t ino, const char *name,
 			const char *value, size_t size, int flags)
 {
 	char procname[64];
+	struct lo_data *lo = lo_data(req);
 	struct lo_inode *inode = lo_inode(req, ino);
 	ssize_t ret;
 	int saverr;
@@ -1188,6 +1376,8 @@ static void lo_setxattr(fuse_req_t req, fuse_ino_t ino, const char *name,
 	ret = setxattr(procname, name, value, size, flags);
 	saverr = ret == -1 ? errno : 0;
 
+	if (!saverr)
+		update_version(lo, inode);
 out:
 	fuse_reply_err(req, saverr);
 }
@@ -1195,6 +1385,7 @@ out:
 static void lo_removexattr(fuse_req_t req, fuse_ino_t ino, const char *name)
 {
 	char procname[64];
+	struct lo_data *lo = lo_data(req);
 	struct lo_inode *inode = lo_inode(req, ino);
 	ssize_t ret;
 	int saverr;
@@ -1219,6 +1410,8 @@ static void lo_removexattr(fuse_req_t req, fuse_ino_t ino, const char *name)
 	ret = removexattr(procname, name);
 	saverr = ret == -1 ? errno : 0;
 
+	if (!saverr)
+		update_version(lo, inode);
 out:
 	fuse_reply_err(req, saverr);
 }
@@ -1332,6 +1525,107 @@ static struct fuse_lowlevel_ops lo_oper = {
         .removemapping  = lo_removemapping,
 };
 
+static void *ireg_do(void *data)
+{
+	struct lo_data *lo = data;
+	int res;
+	char buf[100];
+	struct srv_msg reply;
+	struct msgreply *rep;
+
+	for (;;) {
+		res = read(lo->ireg_sock, buf, sizeof(buf));
+		if (res <= 0) {
+			if (res == -1)
+				warn("read(lo->ireg_sock, ...)");
+			else
+				warnx("disconnected from ireg");
+			return NULL;
+		}
+		if (res != sizeof(reply)) {
+			warnx("bad size message: %i", res);
+			continue;
+		}
+
+		memcpy(&reply, buf, sizeof(reply));
+		if (reply.op != SRV_VERSION) {
+			warn("bad reply to IREG_GET: %i", reply.op);
+			continue;
+		}
+
+		rep = (struct msgreply *) reply.handle;
+		rep->inode->version_offset = reply.version.offset;
+		rep->inode->ireg_refid = reply.version.refid;
+		sem_post(&rep->ready);
+	}
+}
+
+static void setup_shared_versions(struct lo_data *lo)
+{
+	int fd, sock, res;
+	const char *version_path = "/dev/shm/fuse_shared_versions";
+	struct stat stat;
+	struct sockaddr_un name = { .sun_family = AF_UNIX };
+	const char *socket_name = "/tmp/ireg.sock";
+	void *addr;
+
+	lo->ireg_sock = -1;
+	if (lo->cache != CACHE_SHARED)
+		return;
+
+	sock = socket(AF_UNIX, SOCK_SEQPACKET, 0);
+	if (sock == -1)
+		err(1, "socket(AF_UNIX, SOCK_SEQPACKET, 0)");
+
+	strncpy(name.sun_path, socket_name, sizeof(name.sun_path) - 1);
+
+	res = connect(sock, (const struct sockaddr *) &name,
+		      sizeof(struct sockaddr_un));
+	if (res == -1) {
+		warn("connect to ireg");
+		close(sock);
+		lo->ireg_sock = -1;
+		return;
+	}
+
+	lo->ireg_sock = sock;
+
+	fd = open(version_path, O_RDWR);
+	if (sock == -1)
+		err(1, "open(%s, O_RDWR)", version_path);
+
+	res = fstat(fd, &stat);
+	if (res == -1)
+		err(1, "fstat(%i, &stat)", fd);
+
+	lo->version_table_size = stat.st_size / sizeof(lo->version_table[0]);
+
+	addr = mmap(NULL, stat.st_size, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+	if (addr == MAP_FAILED)
+		err(1, "mmap(NULL, %li, PROT_READ | PROT_WRITE, MAP_SHARED, %i, 0", stat.st_size, fd);
+
+	lo->version_table = addr;
+}
+
+static void setup_root(struct lo_data *lo, struct lo_inode *root)
+{
+	int fd, res;
+	struct stat stat;
+
+	fd = open(lo->source, O_PATH);
+	if (fd == -1)
+		err(1, "open(%s, O_PATH)", lo->source);
+
+	res = fstatat(fd, "", &stat, AT_EMPTY_PATH | AT_SYMLINK_NOFOLLOW);
+	if (res == -1)
+		err(1, "fstatat(%s)", lo->source);
+
+	root->fd = fd;
+	root->ino = stat.st_ino;
+	root->dev = stat.st_dev;
+	root->refcount = 2;
+}
+
 int main(int argc, char *argv[])
 {
 	struct fuse_args args = FUSE_ARGS_INIT(argc, argv);
@@ -1375,7 +1669,6 @@ int main(int argc, char *argv[])
 		return 1;
 
 	lo.debug = opts.debug;
-	lo.root.refcount = 2;
 	if (lo.source) {
 		struct stat stat;
 		int res;
@@ -1400,6 +1693,7 @@ int main(int argc, char *argv[])
 			lo.timeout = 1.0;
 			break;
 
+		case CACHE_SHARED:
 		case CACHE_ALWAYS:
 			lo.timeout = 86400.0;
 			break;
@@ -1408,9 +1702,9 @@ int main(int argc, char *argv[])
 		errx(1, "timeout is negative (%lf)", lo.timeout);
 	}
 
-	lo.root.fd = open(lo.source, O_PATH);
-	if (lo.root.fd == -1)
-		err(1, "open(\"%s\", O_PATH)", lo.source);
+	setup_shared_versions(&lo);
+
+	setup_root(&lo, &lo.root);
 
 	se = fuse_session_new(&args, &lo_oper, sizeof(lo_oper), &lo);
 	if (se == NULL)
@@ -1424,12 +1718,26 @@ int main(int argc, char *argv[])
 
 	fuse_daemonize(opts.foreground);
 
+	if (lo.ireg_sock != -1) {
+		pthread_t ireg_thread;
+
+		ret = pthread_create(&ireg_thread, NULL, ireg_do, &lo);
+		if (ret) {
+			warnx("pthread_create: %s", strerror(ret));
+			ret = 1;
+			goto err_out4;
+		}
+
+		get_shared(&lo, &lo.root);
+	}
+
 	/* Block until ctrl+c or fusermount -u */
 	if (opts.singlethread)
 		ret = fuse_session_loop(se);
 	else
 		ret = fuse_session_loop_mt(se, opts.clone_fd);
 
+err_out4:
 	fuse_session_unmount(se);
 err_out3:
 	fuse_remove_signal_handlers(se);
